+++ src/app/organizations/[id]/page.tsx
@@ -1,0 +1,4 @@
+export default function Page({ params }: { params: { slug: string } }) {
+  return <div>My Post: { params.slug }</div>;
+}
+
